+++ include/core/parameters.h
@@ -1166,6 +1166,15 @@ namespace Parameters
     /// prefix for the total volume output in cfd-dem simulation
     std::string phase_volumes_output_name;
 
+    /// Enable output of Q-criterion field
+    bool output_q_criterion;
+
+    /// Enable output of vorticity field
+    bool output_vorticity;
+
+    /// Enable output of velocity gradient field
+    bool output_velocity_gradient;
+
     static void
     declare_parameters(ParameterHandler &prm);
     void